backend/requirements.txt
# Project Sentinel - Backend Dependencies
# ========================================
# Python 3.9+ required

# GDELT Data Ingestion
gdelt>=0.1.14            # Python wrapper for GDELT 2.0 API (PyPI: gdelt)
pandas>=2.0.0            # DataFrame manipulation for event data

# HTTP & Retry Logic
requests>=2.31.0         # HTTP client for API calls
tenacity>=8.2.0          # Retry decorator for network resilience

# RSS/News Ingestion
feedparser>=6.0.0        # RSS/Atom feed parsing

# Telegram MTProto (User Client)
telethon>=1.34.0         # MTProto client for Telegram scraping
python-dotenv>=1.0.0     # Load .env configuration

# JSON Processing
python-dateutil>=2.8.2   # Date parsing and timezone handling
